# docs/openapi/openapi-v2.yaml
openapi: 3.1.0
info:
  title: CircleHaven — Trust Circle API v2 (E2EE X25519/Ed25519)
  version: "2.0.0"
  description: Les corps Auth sont limités à 16 Kio et les corps Messaging à 256 Kio. Les objets d'entrée refusent les propriétés inconnues.
servers:
  - url: https://trust-circle.kavalek.fr
security:
  - bearerAuth: []
    deviceIdHeader: []
    deviceKeyVersionHeader: []
    deviceProofHeader: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token EdDSA/Ed25519 uniquement, audience trust-circle-api et typ=access.
    refreshAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token uniquement, audience trust-circle-auth et typ=refresh.
    deviceIdHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Id
      description: UUID de l'appareil du compte qui produit la preuve d'accès.
    deviceKeyVersionHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Key-Version
      description: Version entière positive de la clé d'identité Ed25519 de l'appareil.
    deviceProofHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Proof
      description: Signature Ed25519 Base64 canonique de la transcription d'accès de 89 octets liée au jti de l'access token. Les trois en-têtes appareil sont obligatoires ensemble sur les routes Messaging, sauf enrôlement initial.
  schemas:
    DeviceKey:
      type: object
      additionalProperties: false
      required: [userId, deviceId, pk_sig, pk_kem, key_version, status]
      properties:
        userId: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        pk_sig: { type: string, minLength: 44, maxLength: 44, pattern: '^[A-Za-z0-9+/]{43}=$', description: Base64 canonique de 32 octets }
        pk_kem: { type: string, minLength: 44, maxLength: 44, pattern: '^[A-Za-z0-9+/]{43}=$', description: Base64 canonique de 32 octets }
        key_version: { type: integer, minimum: 1, maximum: 4294967295 }
        status: { type: string, enum: [active, superseded, revoked] }
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time, nullable: true }
...
paths:
...
  /keys/group/{groupId}:
    parameters:
      - in: path
        name: groupId
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [keys]
      summary: Répertoire signé des clés courantes et historiques du cercle
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DeviceKey' }
        "403": { description: Cercle inaccessible. }
  /keys/group/{groupId}/my-devices:
    get:
      tags: [keys]
      summary: Lister ses versions de clés actives, remplacées et révoquées dans un cercle membre
      parameters:
        - in: path
          name: groupId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
        "403": { description: Cercle inaccessible. }
  /keys/group/{groupId}/devices:
    post:
      tags: [keys]
      summary: Publier ou faire tourner ses propres clés versionnées et signées
      parameters:
        - in: path
          name: groupId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [deviceId, pk_sig, pk_kem, key_version, identityKeyVersion, bindingSignature]
              properties:
                deviceId: { type: string, format: uuid }
                pk_sig: { type: string, description: Base64 canonique de la clé Ed25519 brute de 32 octets }
                pk_kem: { type: string, description: Base64 canonique de la clé X25519 brute de 32 octets }
                key_version: { type: integer, minimum: 1, maximum: 4294967295 }
                identityKeyVersion: { type: integer, minimum: 1, maximum: 4294967295 }
                bindingSignature: { type: string, description: Signature Ed25519 Base64 canonique de la transcription de liaison de 152 octets }
      responses:
        "201":
          description: Publication initiale, rejeu idempotent ou rotation séquentielle acceptée
          content:
            application/json:
              schema:
                type: object
                required: [ok, keyVersion, rotated]
                properties:
                  ok: { type: boolean }
                  keyVersion: { type: integer, minimum: 1 }
                  rotated: { type: boolean }
        "403": { description: Cercle inaccessible, liaison invalide ou appareil non actif/révoqué. }
        "409": { description: Version obsolète, conflictuelle ou non immédiatement successive. }
  /keys/group/{groupId}/devices/{deviceId}:
    delete:
      tags: [keys]
      deprecated: true
      summary: Ancienne révocation par cercle neutralisée
      parameters:
        - in: path
          name: groupId
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: deviceId
          required: true
          schema: { type: string }
      responses:
        "409": { description: Utiliser la décision signée revoke du registre global d'appareils. }
...